pyproject.toml
[project]
name = "openresearch-graph-starter"
version = "0.1.0"
description = "Embeddable research skill graph starter"
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["packages"]
